requirements.txt
autopep8==1.7.0
brotlipy==0.7.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
clic==0.1.3
colorama==0.4.5
cryptography==38.0.1
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.15.0
pyOpenSSL==22.0.0
PySocks==1.7.1
requests==2.28.1
six==1.16.0
toml==0.10.2
urllib3==1.26.12
Werkzeug==2.2.2
win-inet-pton==1.1.0
wincertstore==0.2
